+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## 0.0.1
+## 0.0.2-dev
 
-- Initial version, created by Stagehand
+- Separated linear algebra into its own package
+
+## 0.0.1-dev
+
+- `OLS` implementation
+- `RegressionResult` implementation